Relocated FrontendTypes and fixed imports
