+++ src/app/components/StateRoute/History.tsx
@@ -5,7 +5,7 @@ import React, { useEffect } from 'react';
 // formatting findDiff return data to show the changes with colors, aligns with actions.tsx
 import { diff, formatters } from 'jsondiffpatch';
 import * as d3 from 'd3';
-import { DefaultMargin } from '../../components/FrontendTypes';
+import { DefaultMargin } from '../../FrontendTypes';
 import { changeView, changeSlider, setCurrentTabInApp } from '../../actions/actions';
 import { useStoreContext } from '../../store';
